Address the second CodeQL pass on the Force Generator
Removes javadoc blocks orphaned by earlier edits (they documented methods
that had moved and carried @param tags for parameters that no longer exist),
and widens a test counter to long so a long count is not narrowed on
compound assignment.

+++ megamek/unittests/megamek/client/ratgenerator/ClanEnhancedImagingAugmentorTest.java
@@ -133,7 +133,7 @@ class ClanEnhancedImagingAugmentorTest {
     /** Around one warrior in twenty, not most of them and not none. */
     @Test
     void onlyASmallShareOfWarriorsCarryTheImplant() {
-        int implanted = 0;
+        long implanted = 0;
         int total = 0;
         for (int sample = 0; sample < SAMPLES; sample++) {
             List<Entity> warriors = new ArrayList<>();
